requirements.txt
asgiref==3.8.1
astroid==3.3.8
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.42.2
dill==0.3.9
Django==3.2.25
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
idna==3.10
isort==6.0.0
mccabe==0.7.0
Pillow==10.1.0
platformdirs==4.3.6
pylint==3.3.4
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pytz==2025.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.2
urllib3==2.3.0
